Moved code between the backwards compatibility and the new cluster object headers
