+++ src/app/clusters/electrical-power-measurement-server/electrical-power-measurement-server.h
@@ -17,89 +17,20 @@
  */
 #pragma once
 
-#include <lib/core/Optional.h>
 
-#include <app-common/zap-generated/cluster-objects.h>
+#include "ElectricalPowerMeasurementCluster.h"
 #include <app/AttributeAccessInterface.h>
-#include <protocols/interaction_model/StatusCode.h>
 
 namespace chip {
 namespace app {
 namespace Clusters {
 namespace ElectricalPowerMeasurement {
 
-class Delegate
-{
-public:
-    virtual ~Delegate() = default;
-
-    void SetEndpointId(EndpointId aEndpoint) { mEndpointId = aEndpoint; }
-
-    virtual PowerModeEnum GetPowerMode()          = 0;
-    virtual uint8_t GetNumberOfMeasurementTypes() = 0;
-
-    /* These functions are called by the ReadAttribute handler to iterate through lists
-     * The cluster server will call Start<Type>Read to allow the delegate to create a temporary
-     * lock on the data.
-     * The delegate is expected to not change these values once Start<Type>Read has been called
-     * until the End<Type>Read() has been called (e.g. releasing a lock on the data)
-     */
-    virtual CHIP_ERROR StartAccuracyRead()                                                     = 0;
-    virtual CHIP_ERROR GetAccuracyByIndex(uint8_t, Structs::MeasurementAccuracyStruct::Type &) = 0;
-    virtual CHIP_ERROR EndAccuracyRead()                                                       = 0;
-
-    virtual CHIP_ERROR StartRangesRead()                                                 = 0;
-    virtual CHIP_ERROR GetRangeByIndex(uint8_t, Structs::MeasurementRangeStruct::Type &) = 0;
-    virtual CHIP_ERROR EndRangesRead()                                                   = 0;
-
-    virtual CHIP_ERROR StartHarmonicCurrentsRead()                                                     = 0;
-    virtual CHIP_ERROR GetHarmonicCurrentsByIndex(uint8_t, Structs::HarmonicMeasurementStruct::Type &) = 0;
-    virtual CHIP_ERROR EndHarmonicCurrentsRead()                                                       = 0;
-
-    virtual CHIP_ERROR StartHarmonicPhasesRead()                                                     = 0;
-    virtual CHIP_ERROR GetHarmonicPhasesByIndex(uint8_t, Structs::HarmonicMeasurementStruct::Type &) = 0;
-    virtual CHIP_ERROR EndHarmonicPhasesRead()                                                       = 0;
-
-    virtual DataModel::Nullable<int64_t> GetVoltage()         = 0;
-    virtual DataModel::Nullable<int64_t> GetActiveCurrent()   = 0;
-    virtual DataModel::Nullable<int64_t> GetReactiveCurrent() = 0;
-    virtual DataModel::Nullable<int64_t> GetApparentCurrent() = 0;
-    virtual DataModel::Nullable<int64_t> GetActivePower()     = 0;
-    virtual DataModel::Nullable<int64_t> GetReactivePower()   = 0;
-    virtual DataModel::Nullable<int64_t> GetApparentPower()   = 0;
-    virtual DataModel::Nullable<int64_t> GetRMSVoltage()      = 0;
-    virtual DataModel::Nullable<int64_t> GetRMSCurrent()      = 0;
-    virtual DataModel::Nullable<int64_t> GetRMSPower()        = 0;
-    virtual DataModel::Nullable<int64_t> GetFrequency()       = 0;
-    virtual DataModel::Nullable<int64_t> GetPowerFactor()     = 0;
-    virtual DataModel::Nullable<int64_t> GetNeutralCurrent()  = 0;
-
-protected:
-    EndpointId mEndpointId = 0;
-};
-
-enum class OptionalAttributes : uint32_t
-{
-    kOptionalAttributeRanges          = 0x1,
-    kOptionalAttributeVoltage         = 0x2,
-    kOptionalAttributeActiveCurrent   = 0x4,
-    kOptionalAttributeReactiveCurrent = 0x8,
-    kOptionalAttributeApparentCurrent = 0x10,
-    kOptionalAttributeReactivePower   = 0x20,
-    kOptionalAttributeApparentPower   = 0x40,
-    kOptionalAttributeRMSVoltage      = 0x80,
-    kOptionalAttributeRMSCurrent      = 0x100,
-    kOptionalAttributeRMSPower        = 0x200,
-    kOptionalAttributeFrequency       = 0x400,
-    kOptionalAttributePowerFactor     = 0x800,
-    kOptionalAttributeNeutralCurrent  = 0x1000,
-};
-
 class Instance : public AttributeAccessInterface
 {
 public:
     Instance(EndpointId aEndpointId, Delegate & aDelegate, BitMask<Feature> aFeature,
-             BitMask<OptionalAttributes> aOptionalAttributes) :
+                BitMask<OptionalAttributes> aOptionalAttributes) :
         AttributeAccessInterface(MakeOptional(aEndpointId), Id),
         mDelegate(aDelegate), mFeature(aFeature), mOptionalAttrs(aOptionalAttributes)
     {
